#326 Move edge prototypes to DiagramType
